Update patch set 4
Patch Set 4:

This fixes LTE on verizon. Without this, any change to the network mode prevents a mode of 7 being set again, because it isn't recognized as an LTE mode. It's instead listed as NT_MODE_GLOBAL. Would a better solution be to go through RIL and where the consts are defined and map those appropriately?

Patch-set: 4