Update patch set 4
Patch Set 4:

Could we just add an if statement around this to use the fix for devices that it's needed for? Actually, I think the change to the first switch statement can't have any repercussions, and maybe for the switch statement changed in UpdateEnabledNetworksSummary there could be an if statement before the switch to check if NetworkMode is Phone.NT_MODE_GLOBAL and check if it should be changed to Phone.NT_MODE_LTE_CDMA_AND_EVDO if the device is bugged. Also, I think the statements for Phone.NT_MODE_WCDMA_PREF and Phone.NT_MODE_CDMA could be merged while we're at it.

Patch-set: 4