Update patch set 4
Patch Set 4:

I can't even fathom how that is even remotely related to the issue this fix addresses, are you sure it's not something else or that you're not still using this? To reproduce the issue that this fixes, you have to go into the mobile networks menu and switch from LTE (the default on a new flash) to 3G. When you try to turn LTE back on, it won't work anymore. Just making sure you've done that...

Patch-set: 4