Update patch set 7
Patch Set 7:

galaxys2 is a no-op, too. Nothing in its HAL makes use of this (no references to either noise_suppression or dualmic_enabled)

The only audio HAL I can find with active use of a noise suppression argument is the i9300's (device/samsung/i9300/audio/audio_hw.c), everything else just defaults to enabling it and has no toggle mechanisms. I'm not sure this is worth doing for the sake of a single device, especially since the implementation is non-obvious and other devices appear to be enabling it, even if it does nothing. It looks more of a placebo than an actual feature.

Patch-set: 7