Update patch set 4
Patch Set 4:

Would it be possible to add a opencnam professional key to it?  I find the lookups are more accurate under the professional tier service.  I hacked up this app: http://www.integralblue.com/callerid-for-android/
And now, using my own server, it displays the opencnam results, as well as a spam score, from whocalled.us. 

Anyway, even the possibility of selecting a custom URL that returned results via JSON or something could be useful.

Patch-set: 4