Update patch set 1
Patch Set 1:

At first boot or when a sim is added you get this dialog.

http://i158.photobucket.com/albums/t109/dhacker29/Screenshot_1970-01-03-01-38-30.png

Currently it does nothing, so if you are in SetupWizard you can't activate the sim.  This path links the ok to the settings.

http://i158.photobucket.com/albums/t109/dhacker29/Screenshot_1970-01-03-01-38-38.png

Not sure if it's possible or necessary to go to the specific preference item?

http://i158.photobucket.com/albums/t109/dhacker29/Screenshot_1970-01-03-01-38-47.png

Patch-set: 1