Update patch set 2
Patch Set 2:

I'm not convinced of using a (assumed no-op) activity as trampoline for another one. Either the activity is a no-op, in which case kill it with fire, or it's not, in which case this patch is wrong. Either way, in its current state the patch looks like a band aid.

Patch-set: 2