Update patch set 2
Patch Set 2:

Well it's not really a no-op it's just causing the button to be automatically clicked. I honestly can't figure out how the previous mobile network settings intent triggered this screen directly. But yes it's a bandaid to fix a bandaid/hack that moved this to main settings. Short of rewriting the MSimMobileNetworkSettings I'm not sure how to make this perform as intended.

Patch-set: 2