Update patch set 1
Patch Set 1:

Any chance we can get this in for M6?  Please keep in mind both that,

This patch has been sitting here for over a month.  It's not a particularly complicated patch either.  It shouldn't take this long to review.

This patch is a _straight port_ of one we included in CM 9 and CM 10.x.  It's not new, it's been around for a long time.

In fact, the only reason I had to reupload the patch at all is due to the CDMA system select code being moved upstream from android_packages_apps_Phone to android_packages_services_Telephony, otherwise it would already be in CM 11.

Patch-set: 1