Update patch set 1
Patch Set 1:

There is no values changed here. it is just redirecting the hardcoded values to xml file which makes it a LOT more flexible than the hard coded data in switch statement. I think the one reason why they did it that way is speed and performance instead of reading the xml.  This does nothing except change the data input/output it is coming from. this should be merged without any hesitation, if you read mkasick comment in the code.

Patch-set: 1