Update patch set 8
Patch Set 8: Code-Review+1

https://dl.dropboxusercontent.com/u/28823036/G/Screenshot_2014-04-28-22-34-11.png

I am not sure I understand how to do this without a popup.  You have the option in MSIM Settings to choice a default sim for voice calls or to be prompted for each call. When prompted you have to choice which sim to dial out with.

Patch-set: 8