Update patch set 11
Patch Set 11:

Also after thinking about this. The plan is to add an option to contacts to remember prefered sim per contact like custom ringtone. In stock there is a check box at the bottom of the dialog "Remember my selection"  Not sure how hard that would be to add to the activity picker type view?

Patch-set: 11