Update patch set 3
Abandoned

I've lost interest in this. invisiblek, feel free to resurrect this on your own, but to be honest, it's really only an issue for devices that should be world_phone capable, but cannot use that overlay because of device unification. So, not sure if services_Telephony is the right place to fix this.

Patch-set: 3
Status: abandoned