Update patch set 3
Patch Set 3:

I was thinking of someone who attempts to pickup the latest version of the dialer with and older build of CM, but side loading the google dialer is a good example also.  Agreed that these cases are rare, but again what's the downside of appending new fields at the end?  Seems like a good practice in general and the upside is that it could prevent breaking the dialer completely for one or two users.

Patch-set: 3