Update patch set 3
Patch Set 3:

My point is that we already _are_ incompatible since right the beginning of the 4.4 development cycle and seemingly nobody has noticed. That's why I think it's a pointless exercise to specially care for compatibility when adding new fields, and I think it's more important to keep the logical grouping in that case. I can change it back if you really feel it's important, but IMHO it's not.
(BTW, I don't think there's a single end user who sideloads a Dialer apk of a newer CM build into an older release of CM)

Patch-set: 3