Update patch set 2
Patch Set 2:

I do see what you are saying. For the world phone case, the prefs launch GsmUmtsCallOptions as a separate Preference Activity. The setting exists there. But it is broken. Not appending the sub to the pref. For the non world phone, the pref is removed, added inline, and handled in CallFeatures. I'll fix the broken case and append.

Patch-set: 2