Update patch set 2
Patch Set 2:

Updated the commit message.

This occurs when you're on RIL_RADIO_TECHNOLOGY_HSDPA (9), which gets converted to NETWORK_TYPE_HSDPA (8). After that, the code will think that it is RIL_RADIO_TECHNOLOGY_EVDO_A (8) and try to get the IccRecord for the CDMA family, which will fail because the phone is in the GSM family.

Patch-set: 2
Reviewer: Pawit Pornkitprasan <1381@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0