Update patch set 5
Patch Set 5:

Hey Danny,

It looks like SSNs for VT calls are no longer propagating after this change.  I think it is due to the fact that we are now only registering a listener for SSNs on GSMPhones, whereas before we were registering in TelephonyConnection using  getPhone().registerForSuppServiceNotification().  One of the possible return values for TelephonyConnection.getPhone() is an ImsPhone.  (It took me a while to figure out how this is possible... GSM/CDMA phones delegate to an mImsPhone data member on dial).

Since there are two types of phones that have SSNs, it may make sense to move the registration back to TelephonyConnection.  I will start working on a patch, but wanted to give you a heads up in case you have ideas on it.

Patch-set: 5
Reviewer: Ed Mancebo <10293@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0