Update patch set 2
Patch Set 2:

Which is like ... exactly what we did in the caf/cm-12.0 version of this patch? This was in review for a couple of days, it wasn't ninja merged.

There will always be stuff that slips through review. Sucks, but the code is too complex to cover each and every possible bug.

Patch-set: 2