Update patch set 8
Patch Set 8:

How does all of this help with the fact that MobileNetworkSettings controls the PREFERRED_NETWORK_MODE setting, which is updated by setPreferredNetworkType (e.g. ModemBindingPolicyHandler.java:256) and thus will always be '2G' for the non-data SIM? The goal of those patches is that the UI setting always reflects the user's choice, and not necessarily the current state.

AFAICT, there's only 3 possible solutions for this:
- Don't use PREFERRED_NETWORK_MODE to store the user's choice, but something else (our approach)
- Don't overwrite PREFERRED_NETWORK_MODE after setPreferredNetworkType() succeeded, and treat PREFERRED_NETWORK_MODE like our CONFIGURED_NETWORK_MODE (not sure about the consequences of that)
- Do the 2G fallback for the non-data SIM via a mechanism distinct from setPreferredNetworkType() (best solution, implemented in AOSP RIL v11, but not sure if/when Qualcomm will support it)

AFAICT the LTE toggle is broken for MSIM anyway - it always operates on the first phone (and should probably rather be implemented via the PhoneToggler). What other TelephonyManager API are you referring to?

Patch-set: 8