Update patch set 8
Patch Set 8:

Here's my thought process.  Currently PREFERRED_NETWORK_MODE is overloaded to mean two things:

- It is the network mode that the user picked.  Typically it is picked from the dropdown, but it could be picked from somewhere else like the LTE tile or when the user re-inserts a sim.
- It is the current network mode of each phone.

It is pretty unfortunate that the setting is used for two different things, but I'd be hesitant to change the meaning of PREFERRED_NETWORK_MODE because it is exposed in several places and accessed a lot.  An app can set network mode via telephony manager or by broadcasting to phone toggler--if they think that doing so will change the user's preferred network but we are now using CONFIGURED_NETWORK_MODE, then it breaks.  

The other (more minor) concern is that we could confuse future developers if the UI says preferred network mode but actually set configured network mode, but there is still a setting called preferred network mode that now means something different. 

My proposal is to not change the meaning of preferred network mode.  We can just stash the current network mode when we do the fallback to 2G so that we can restore it when the user switches back to that sim.

I'll take another look at the LTE toggle, if it is broken we should fix it or deprecate it.  The TelephonyManager API I was referring to is TelephonyManager.setPreferredNetworkMode, again the concern being that any future code that calls that method could be broken if we change the meaning of preferred network mode.

Re: L256 of ModemBindingPolicyHandler which is used to restore the network mode for each sim from the subs table.  Afaict that code path is invoked when you insert a sim.  What I am proposing is pretty similar--restore the network mode from a previous state.

Patch-set: 8