Update patch set 2
Patch Set 2:

I suppose we could wrap in a flag if we want to be extra careful.  On the other hand the use case is pretty narrow already (failover to another sub on an emergency call that disconnects with an error code).  And it appears to be broken as is.

Patch-set: 2