Fix: Send Deactivate SMS after Native VVM deactivation
When the system detects the Carrier vvm app, the Native VVM Implementation
is shut down. But, a deactivation SMS is missing to the OMTP server.
Whitout this message, all OMTP notifications (Binaries SMS) are not catch
by the native VVM implementation and not by the Carrier VVM App (if it is
not yet launched).This use case is relevant in case of the Carrier VVM Ap
p is pushed through the Play Auto Install mechanism or any others. The
carrier application could be never launched.

Change-Id: If412e73fe582e517ef44a183ef028d347b44b82a
Signed-off-by: Alexis MARCOU <alexis.marcou@orange.com>