Update patch set 4
Patch Set 4:

why did we add this back, this will break the logic that we put. The idea is that if there is no extphone service then we want to fall through and go to the subid logic below it. By putting the return statement it will not go to the subid logic below after the exception.

Patch-set: 4
Reviewer: Deepak Kundra <13513@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0