Defer VVM activation/deactivation until device is unlocked.
VVM activation/deactivation requires voicemail status table update and
sending SMS which is not available if the device is locked. The
activation/deactivation also requires the subId of the SIM card, which
is enumerated when the SIM loads, with the possibility of the device
still being locked.

This CL checks if the device is locked when SimChangeReceiver is
triggered. If so, the subId will be stored by DeferredSimChangeProcessor
and resent to SimChangeReceiver after the device is unlocked.

VVM SYNC SMS could also be received when the device is locked. The SMS
will simply be ignored, as a full sync will always be performed when the
device in unlocked.

Bug:27534089
Bug:26261432
Change-Id: I9a966e5b1b880a764e4ca6b18e77796a88496043