Support hiding manage IMS conference call button
Check carrier config in TelecomAccountRegistry when registering
PhoneAccounts. When creating new TelephonyConnections, set a propery
on the connection to indicate whether it supports managing IMS
conference call.
This is then used in the conference to determine whether the
conference can manage the children calls.

Test: manual - checked the manage ims conference call button
Bug: 28729600
Change-Id: I1000830a85814150658b62ea9add72bfb83bbba0
Merged-In: I1000830a85814150658b62ea9add72bfb83bbba0