Add telephony side visual voicemail config
Carrier specific visual voicemail configs used to be managed by
CarrierConfigManager. Adding a new config will require public API
changes, which we are prevented from doing so. Also some settings are
better not to be made public as it will complicate the settings to much.

For example, some carrier provides direct SSL connection to their IMAP
server. This port used to be hard coded, and we cannot add it to the
public API until next SDK release. Some carrier is also known to have
faulty implementation in their IMAP server, so we need to avoid using
some command. This config which is for some specific situation is not
suitable for public API.

This CL add another layer of config, TelephonyVvmConfig.
TelephonyVvmConfigManager will load the configs from a xml file, and
provide a suitable config based on MCC MNC. OmtpVvmCarrierConfigHelper
will still prioritize configs provided by CarrierConfigManager, but will
use the values from telephony if CarrierConfigManager does not provide
the config or the config is not public.

+ Mutliple carrier package name support
+ Proguard flags to prevent VisibleForTesting from being removed.

Change-Id: I9be7f4c8b45100a4ac9a4d6bca1842f2e244a99d