Update patch set 2
Patch Set 2: Code-Review-1

I still don't understand why we do this for resources which are pulled from an upstream source, in this case CAF.  Does it solve an actual problem?  Telephony is already hard enough to merge and if this isn't anything other than OCD then let's not make the next merge even more hellish.

Patch-set: 2