Update patch set 2
Patch Set 2:

The Translation system policy makes core assumptions around what files are translated for AOSP supported languages, and CM-additional languages. 

AOSP strings (with each upstream release) come with translations, and therefore only the CM-additional languages are configured to translate them.

https://github.com/CyanogenMod/cm_crowdin/blob/master/config/cm-13.0_aosp.yaml#L492

For our translations purposes, with the goals of: limiting strings conflicts with our additions vs AOSP, porting (identifying ours vs AOSP vs CAF), and again on how the Translation system is configured based on the above, the 'strings.xml' file here isn't available for translators to work on. To facilitate translation work, these strings are moved to cm_strings or qti_strings (qcom_strings). 

https://github.com/CyanogenMod/cm_crowdin/blob/master/config/cm-13.0.yaml

tl;dr Because we want to translate this and not break the tool.

Patch-set: 2