Update patch set 3
Patch Set 3:

As Abhisek mentioned, we move the strings in order to be able to translate them. We used to have a workaround in the sync script which moved the CAF string additions to a new temporary file, but it was not worth the trouble. We also modify almost all the strings because they suck, so it should still be hard to merge even if we didn't move them.

Patch-set: 3