Update patch set 5
Patch Set 5:

The final fix is to use the phoneId based on which RIL is initialized. We were trying to dial on SUB0 while actually SUB1 was active and vice versa. So using ModemUuid you can be sure that the right SUB is dialled on.

4847 07-01 20:37:39.364  4357  4357 I Telephony: TelephonyConnectionService: pickBestPhoneForEmergencyCall, return primary stack phone
4848 07-01 20:37:39.364  4357  4357 I Telephony: TelephonyConnectionService: Logical Modem id: 1 phoneId: 0
4849 07-01 20:37:39.364  4357  4357 I Telephony: TelephonyConnectionService: Logical Modem id: 0 phoneId: 1
4850 07-01 20:37:39.364  4357  4357 I Telephony: TelephonyConnectionService: Primay Stack phone id: 1 selected

Patch-set: 5
Reviewer: Deepak Kundra <13513@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>