Update patch set 2
Patch Set 2:

(2 comments)

Some minor things left. Did you see any issues with naming the defintion EXTRA_INITIAL_SLOT_TAB (or something among those lines)? Because that's what it actually does; in my POV the fact that it's due to a roaming notification is a detail up to the caller.

Patch-set: 2