IMS Conference fixes related to RemoteConnectionServices.
1. In ImsConferenceController, adding the children of an IMSConference
to the list of conferencable connections.  This is necessary because the
RemoteConnectionService API does not support specifying a list of
Conferenceables for a RemoteConnection.  Where Conferenceables are either
Connections or Conferences.  It only supports specifying a list of
conferenceable connections.  This is problematic because for an IMS
conference call, a standlone Connection is technically conferenceable with
the ImsConference.  As a result, the RemoteConnetionServices API is not
able to support conferencing a standalone Connection into a Conference.
To fix this temporarily, I'm adding the child ImsConference participant
connections to the list of conferenceables for the standalone Conference.

2. In ImsConference, there was an assumption made in "onMerge" that the
Connection merged into the conference was a TelephonyConnection.  Because
of the issues in (1), we can't make this assumption.  Since the cast
was only there to get at the phone, opted to just use the getPHone()
method in the conf host connetion instead.

Bug: 31464792
Change-Id: I5ee13dc71e566831c02d5ced18e0edb0f8b7ac4b