Fix merge button is shown after reaching maximum conference call size
After reaching maximum conference call size allowed, the conferenceable
connection size is zero, the conference controller does not reset
conferenable connection for the conferenced connection so that can merge
the call due to have the capbility.

To avoid it, assgin empty conferenceable connection for the conferenced
connection after it is full.

Change-Id: I035951a982181396d538519a6e15ed4836a6b4bd