Update patch set 1
Patch Set 1:

Just in case I'm not clear: the above linked piece of code adjusts PREFERRED_NETWORK_MODE without the user instructing to do so, effectively killing the user preference. PREFERRED_NETWORK_MODE is (and, AFAICT, always has been) used as 'state a given SIM is supposed to be in at a given time'. USER_NETWORK_MODE, on the other hand, is 'highest mode the user desires a given SIM to be in'. In the DSDS case, where SIM2 is GSM only, those two obviously differ.

Patch-set: 1