Update patch set 1
Patch Set 1:

Ok, I was not a aware of that. I only saw that there used to be a single PREFERRED_NETWORK_MODE and, given the commit message, I assumed that USER_NETWORK_MODE was added because of that.

Then I need to update the code and make it work properly with USER_NETWORK_MODE.

Patch-set: 1