Update patch set 1
Patch Set 1:

Danny, I have some doubts regarding the use of USER_NETWORK_MODE.

I wanted to update packages/services/Telephony/MobileNetworkSettings.java so that it also changes USER_NETWORK_MODE, but I'm not sure how I should do that. Right now there's nothing that updates USER_NETWORK_MODE if I saw correctly.

MobileNetworkSettings is written to work with PREFERRED_NETWORK_MODE (obviously). This means that it constantly shows the current network mode given that its value is constantly updated by the framework. This may or may not reflect the user preference. In addition to that, if the user tries to set a mode that gets rejected by the modem, the preference is changed automatically to the mode returned by the modem. So in some cases is not even possible to change the preference.

To have MobileNetworkSettings show the actual preference, I should remove all the references to PREFERRED_NETWORK_MODE and only rely on USER_NETWORK_MODE. This also means that I should ignore the value reported by the modem. We shouldn't even use PREFERRED_NETWORK_MODE as fallback value, or we might show a preference that won't stick, which is confusing IMO.

So I was wondering if we could just change the framework not to have have PREFERRED_NETWORK_MODE changed automatically and drop USER_NETWORK_MODE. I actually have to see if this is feasible, but maybe you already know it.

What do you think about this?

Patch-set: 1