Update patch set 1
Patch Set 1:

Umm, maybe it's better to keep using USER_NETWORK_MODE to avoid major conflicts with future merges. We would have to track every change that has anything to do with PREFERRED_NETWORK_MODE and some additions can go unnoticed.

If we keep USER_NETWORK_MODE, I think we should to modify MobileNetworkSettings, because I think it's confusing for the user as it is.

Patch-set: 1