Update patch set 2
Patch Set 2:

Brandon, the two switch statements are actually different. I should look at the history to know why, but that's how it is.

FYI the network mode preference is completely broken as of now. This is because the framework ignores USER_NETWORK_MODE only if it's -1, which is not true after a data wipe (see https://review.lineageos.org/#/c/156700/) or if you upgrade from cm-13.0 after you changed the preference at least once.

Patch-set: 2