Temporarily use broadcast instead of VisualVoicemailService
Temporarily use a broadcast to notify dialer VVM events instead of
using the VisualVoicemailService.
b/35766990 The VisualVoicemailService is undergoing API changes.
The dialer is in a different repository so it can not be updated in sync
with android SDK. It is also hard to make a manifest service to
work in the intermittent state.

Bug: 35766990
Test: manual smoke test.
Change-Id: I25f13feb468628928e407c746e61a31e4da44c39