Update patch set 2
Patch Set 2: Code-Review-1

This fixes the sim settings crash. But it doesn't make any sense to me: if phonefeatures is NOT installed call barring should be removed. Nowmwe have the situation, that if phonefeatures is not installed button_call_barring_expand is null, so with this patch call baring is NOT removed, because button_call_barring_expand is null, because phonefeatures is not installed... :( Since this means call barring is NOT removed when it should be removedmit makes much more sense to revert the commit, which introduced the crash.

Patch-set: 2
Reviewer: Ingo Prochaska <12847@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=-1