Update patch set 1
Patch Set 1:

> By devices without SIM cards, do you mean devices which do not have
 > cell radios? Those devices shouldn't be showing this menu.

If there isn't an active subscription in general. I tried to fix the same issue some time ago and did this: https://review.lineageos.org/#/c/163770/ (same as when you enable Airplane mode).

This solution is definitely simpler, we added the code that throws the exception.

If I'm not wrong stock Android simply opens a page with no usable option, which should be what you would get using this patch.

Patch-set: 1