Implement dispatching of resolution UI.
Create a subclass of the dispatcher for resolution intents. This is
needed because the regular dispatcher is accessible to any app so that
it can resolve public intents; however, apps should not be able to
start the resoution dispatcher directly as one of the dispatcher's
argments is the operation to perform next - if apps could spoof this,
they could trick the platform into thinking the caller was somebody
else.

Bug: 33075886
Test: Unit tests and E2E verification with dummy LPA
Change-Id: Ife30de87af7d275424bc254002c426a13d66178f