Added unmetered use only data connection support
If a data network is brought up while data is turned off because
the request APN is unmetered, we mark this data network
as NET_CAPABILITY_NOT_RESTRICTED and will only advertise
the unmetered APN capabilities for this data network. When
data enable condition changes, this data network will be torn down
so a full capable data network can be re-established.

Test: Unit tests
bug: 30936226
Change-Id: Ie138c2767d549294741949ddfa8e21f2493e50ea