Set/read VM vibration and ringtone from default phone if SIM is absent
Dialer voicemail settings use TelecomManager.
getDefaultOutgoingPhoneAccount() to get the phone account handle in
single SIM scenarios. If the SIM is absent, the returned handle will
be invalid(does not correspond to a phone object). Before this CL,
telephony will throw an exception while trying to modify voicemail
settings with an invalid phone account handle.

After this CL, the settings will be applied to the default phone
instead, which correspond to INVALID_SUBSCRIPTION_ID and is
consistent with previous behavior (same phone object when the
telephony voicemail settings is created without a SIM).

Bug: 37161918
Change-Id: I6336ce076d0f38db14c838cb2ac78f5494e28a3e
Fixes: 37161918
Test: manual - toggle voicemail vibration settings without a SIM.