Implement getVisualVoicemailSettings()
This CL allow the system dialer to migrate voicemail settings from
telephony. The following values are provided:

"is_enabled": whether VVM is enabled before. The key will be
missing if the user never modified this setting and the default
should be used.

"default_old_pin": the voicemail PIN scrambled by the
provisioning process. For certain carriers the PIN must be set before
the voicemail feature can be used. Telephony had a automatic process
that will set a random PIN for the user and then prompt them to change
it. This must be migrated to the system dialer so the PIN will not be
lost after an upgrade.

Bug: 34093562
Test: N/A - pending dialer side changes to be implemented.

Change-Id: I037d1e7ec48eda7a0db282449b1c492e545e8930