Trigger checkRemovedSim when carrier config changed
VvmSimStateTracker scans all known PhoneAccountHandle when a SIM is
removed to inform the VisualVoicemailService which one is removed.
It rely on SubscriptionManager.getActiveSubscriptionInfoForIccIndex()
to determine if a SIM is still present.

The SubscriptionInfo might not be updated yet when
ACTION_SIM_STATE_CHANGED is received. In this CL, the same scan will
be triggered when the ACTION_CARRIER_CONFIG_CHANGED on an invalid
subId.

Bug: 37286583
Change-Id: I0f236527f2d4f8dae0f870f90635e0647f5fb3a8
Fixes: 37286583
Test: manual - remove SIM