Add APM check to EmergencyCallHelper
There is a race condition between APM on and
the radio being turned off. If APM is on and
the radio hasn't been turned off yet, it can
cause the e911 call to fail because the dial
is placed on a radio that is off.

This change turns on EmergencyCallHelper when
either the radio is off or APM is on to eliminate
the race condition.

Test: Manual E911 Tests
Bug: 37960809
Change-Id: I81d269f8a106d1692796788a1974b01162579cc1