Fix to show the network selection no service notification
NotificationMgr uses the CallManager's default phone which could
be IMSPhone if IMSphone registers first. The notification wasn't
shown because of TelephonyCapabilities.supportsNetworkSelection
failed because the phone used is IMSPhone. The fix uses the phone
derived from the subId of ACTION_SERVICE_STATE_CHANGED intent
instead of using CallManager's default phone.

Test: Manually selected a forbidden PLMN and verified the network
      selection no service notifcation.

Bug: 38117310
Change-Id: I62dce24cab349322ee3f4f59ebee1524abb55f91