Declare WRITE_EMBEDDED_SUBSCRIPTIONS in Phone process.
The fix for b/37480432 includes a move of the post-command callback
from a Binder thread to the EuiccConnector handler thread. A side
effect of this is that calls which were previously treated as IPCs
from the LPA (which triggers the callback) are now treated as being
internal to the Phone process. This is more correct, but some
permission checks will begin to fail (like refreshing the list of
subscriptions after a download/delete), so we need to declare this
permission now.

Bug: 37480432
Test: TreeHugger + E2E IPC test
Change-Id: If1eaac585bf28be4955ad8796df3c5d0308c906b