Fix crash issue due to NPE when two calls are merged
NPE occurs when a ViLTE call in hold state and a VoLTE
call in active state are merged. Because an original
connection is cleared once when two calls are merged.
After that, the released original connection is referred
during setting video state.

Fixed to add null check in shouldSetDisableAddCallExtra.

Test: manual - Checked that no crash occurred when a ViLTE call in
hold state and a VoLTE call in active state are merged.
Bug: 66592790

Change-Id: I9101116ba7ef424599ca043d574a30930288d22c