Notify the system dialer when the Carrier VVM app is installed
Before O the telephony VVM app will disable itself if the carrier
VVM app is installed by the user. After O the system dialer which
is in charge of VVM should still respect this behavior.

In this CL a hidden API broadcast will be sent to the system dialer
if the carrier VVM app listed in CarrierConfig is installed.

This is a intermediate solution for O. In P this will either be removed
or made system API.

Bug: 38329114
Test: manual - install carrier VVM app
    install non-VVM app
    install carrier VVM app without a SIM
    install app on carrier without VVM app

Change-Id: I14d3914ceb3abb99a38ad601bb739c15b82f393f