Send IS_REFRESH with ACTION_SHOW_VOICEMAIL_NOTIFICATION
Telephony will resend ACTION_SHOW_VOICEMAIL_NOTIFICATION when the
subscriptions has changed, so the MWI received when the SIM is not
loaded can be posted. For some carriers voicemail count is not
supported so dialer cannot distinguish between a refresh and a new
voicemail, and will post the notification every time (signal lost and
recovery).

Dialer cannot simply use setOnlyAlertOnce() because if it is a new
voicemail the user should be alerted again.

In this CL IS_REFRESH will be sent with
ACTION_SHOW_VOICEMAIL_NOTIFICATION so dialer can know the difference.

Change-Id: I9417e5553bbccfe4feba960c2497fffef48d2fd9
Fixes: 62119940
Test: manual - leave voicemail, toggle airplane mode.