Fix crash issue when user uses multi-window mode on VoicemailSettings
IllegalStateException happens because entries and entryValues are not
set in ListPreference at the timing of onRestoreInstanceState.
Activity recreation occurs due to multi-window mode and
VoicemailSettingsActivity#onRestoreInstanceState is called and then,
ListPreference#onPrepareDialogBuilder is called. ListPreference
requires entries and entryValues but they are still not set. They are
set in VoicemailSettingsActivity.#onResume. But it is called after
VoicemailSettingsActivity#onRestoreInstanceState.

To avoid this crash, add "screenLayout" property in configChanges
attributes of VoicemailSettingsActivity to avoid recreating activity.

Test: manual - Checked Multi-window mode on VoicemailSettings
Bug: 63694457

Change-Id: I2f0d68de25615121a7bd032431f4b153bcb6e8b4