Remove passing of SIP history-info information.
The SuppServiceNotification contains history-info records in the format
described in RFC4244.
Passing of these headers was added in the past, however the functionality
was never used in practice.
The Dialer app displays the most recent history-info record in the UI.
At the time this was added, it was not well understood what the format of
this data would be, and an assumption was made that it would only contain
phone numbers.
In practice, however it will contain history-info records from the SIP
headers.  This data is not in a format suitable for direct display in the
UI.
Since there is no immediate need for this information in the UI, removing
the passing of the data for now.  In the future we will need to implement
a parser for the history-info records and store the data in a format more
amenable for display.

Bug: 63605067
Test: Manual
Change-Id: I735ed4f34324c6e04531a761c80c12d11f74253e