Set default preferred network only when registered on home network.
For certain carriers, we hide "Preferred network type" setting UI and
set to default preferred network type if it's not roaming. Here
we are making sure we do it only when it's registered on home
network. Because during searching, the network type is also "home".

Bug: 63088061
Test: Manual

Merged-In: I121c7712944182fcb3bd1a320616eb57ebe0c0d0
Change-Id: I121c7712944182fcb3bd1a320616eb57ebe0c0d0