PhoneToggler: Update preferred network mode setting
Currently PhoneToggler changes the network mode and the persistent
USER_NETWORK_MODE, but it doesn't update the PREFERRED_NETWORK_MODE
setting. MobileNetworkSettings relies on PREFERRED_NETWORK_MODE
to determine whether it should ask the modem to update the mode
or not. If PREFERRED_NETWORK_MODE is not reporting the current
network mode, MobileNetworkSettings might think that the newly
requested mode is already in used and not request a change.

BUGBASH-898

Change-Id: Icb64ce21c14697006cb5e1327e31d42cee8ede62