Move ACTION_SIP_INCOMING_CALL back to Manifest
The ACTION_SIP_INCOMING_CALL intent receiver along with
all of the other SIP intent receivers were moved to
dynamic registration because background check
was not letting them be received. Since
ACTION_SIP_INCOMING_CALL is an explicit intent, this
caused this intent to not be received anymore and
caused SIP incoming calls to not show InCall UI.

This change moves that intent back to the manifest
so that it can be received.

Bug: 64508600
Test: Manual
Merged-In: I9d1347055c90a49d141da6d34c85ab2a7cb080c1
Change-Id: Ie330f748790fc0470a2b59128258c50e17d4cdf7