Ensure restricted number CDMA call waiting info is passed to Telecom.
The code which notifies Telecom of an incoming call-waiting call over CDMA
made the erroneous assumption that if the phone number provided was empty
that it was not possible to pass the call information to Telecom.

This change specifically checks for the case where the connection's
number presentation is not "allow" and then opts to do the match solely
on the number presentation.

Finally, some more logging is present in all scenarios.

Test: Manual
Bug: 65432905
Change-Id: I13bd319606da450b7bb478f9ce51cde80013b780