Always enabled the Advanced Option under Mobile Network Settings
When the user under pSIM slot without SIM card inserted, the Advanced
option will be disabled. But if the user using eSIM before, the Carrier
option under Advanced is enabled. But the user cannot access it since
the Advanced option is disabled. To solve this problem, we decide always
enable the Advanced option.

Change-Id: I743cb99ffdeb041a0b1b2ea01376abc5b91084d0
Bug: 65292100
Test: E2E