[TelephonyService] Improve hold capability signal
The CAPABILITY_HOLD perpority is not set properly in telephony call(If
there are two top calls, both of them still holdable).

This add a hold tracker to track all telephony connections/conference
and set the holdable state of them. The connections/conference itself
can use the holdable state and the call state to determine if
CAPABILITY_HOLD should be set.

Test: manully test and unit test
Bug: 66949982

Change-Id: I43fde6fcbc047c09ba375c6c8fd5ac374bf7fb70