Update patch set 9
Patch Set 9:

After talking with shareef about how I would have to go about implementing this in a custom RIL class, a few things came to mind:
1. It's insanely complicated to inspect every SMS parcel (since apparently that's how MWIs are sent) and examine them to see if they're an MWI or not, and drop them accordingly. I'll admit I also don't know how to implement such a solution, but that point is moot imo, since that's not the only reason to prefer this solution.
2. O changed a lot of stuff with RIL.java, meaning any logic that was put in there would have to be rewritten/reworked for 15.0, and maybe future android releases as well, but a look at this codepath in 15.0 looks virtually identical to what it is in 14.1.
3. Google actually did something similar right above these code additions, although for a slightly different reason.
4. (and probably my main argument here) This is not a device specific bug. It makes sense to put stuff that's device or vendor specific in device specific RIL classes, but this is a bug on our *entire platform*  for *all Sprint devices*. In addition to the fact that ls990 doesn't need a custom RIL class right now, we would have to duplicate that code across every single device that can be used on Sprint, and any new devices with Sprint support would need to add a custom ril class with those hacks. (also see above for why it would be a royal pain to update that code for new android versions, like Oreo).
In my opinion, if this code works, this is the cleanest way to do it, because it's explicitly applied only for sprint (and sprint MVNOs) via CarrierConfig (instead of relying on users manually enabling a MWI toggle when they're not on Sprint), and it's easy to patch in for existing and upcoming sprint devices, without needing to muck around with messy handling of RIL responses/parcels in custom RIL classes.
As an aside, I was talking to Kelly, and he told me that simply making it not persistent isn't enough, as every time it refreshes network status, it appears again, which is really annoying.
As it is, I'm happy with the code, but will only +1 after I test for regressions.

Patch-set: 9