Fix emergency dialer is unable to be shown in doze
After screen is light by notification due to ambient display feature,
the device is yet in doze, so that shows key guard in doze. When user
clicks the emergency option to launch emergency dialer, its window is
unable to be shown on lock screen.

To fix it, allow turning srceen on, so that window manager can wake up
power, hence, exit doze mode.

Bug: 68678302
Change-Id: I847b95bc281fc0d9ce889aa0d4eb649fc9efe8d9