Fix emergency call is failed without SIM under APM
If phone type is changed after radio power on, the new phone
account handle ID is empty. Without phone account ID to add the
existing connection causes rejecting the request, so that
reconnecting outgoing emergency call is unable to continue.

To fix it, make phone account according to emergency call.

Bug: 68964934
Change-Id: I2be5aa8b69362290c008d6ae1e1a59d7379fe873