Modify the contact picker for call forwarding to directly show contact list
Modified the contact picker preference so that tapping the contact icon
will directly launch the list of contacts instead of going to the system
UI chooser which includes a "Recent document" and "Downloads" option.
The old behavior required the user to tap on the "Contacts" option in the
hamburger menu.  The recent documents and downloads really didn't make
any sense to show.  This has been a bug back to L.

Test: Manually verified that contact list is shown now.
Bug: 27072617
Change-Id: I0a987e50415ad6ab751e16b08c7ef65a2b0e53f9